requirements.txt
torch==2.0.1
numpy==1.24.3
scipy==1.10.1
torchvision==0.15.2
pillow==9.5.0
